добавляет модульную структуру
