+++ js/data.js
@@ -1,51 +1,6 @@
 import {getRandomPositiveInteger} from './util.js';
 import {PHOTOS_COUNT, LikesCount, PhrasesCount, CommentsCount, AvatarCount} from './setup.js';
-
-const PHOTO_DESCRIPTIONS = [
-  'Пляж',
-  'В направлении лета',
-  'Море',
-  'На рассвете',
-  'Кулинарный юмор',
-  'Моя тачка',
-  'Завтрак самурая',
-  'Любимый напиток',
-  'Полет нормальный',
-  'Обувь',
-  'Дорога к морю',
-  'Audi',
-  'Салат',
-  'Котосуши',
-  'Космошузы',
-  'Над облаками',
-  'Распевка',
-  'Ретромобиль',
-  'Теперь можно бегать к холодильнику по ночам',
-  'Пальмы',
-  'Салат',
-  'Закат на море',
-  'Краб',
-  'Вечернее шоу',
-  'Водные процедуры'
-];
-
-const PHRASES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?'
-];
-
-const AUTHORS = [
-  'Мимо проходила',
-  'Кедра Выдра',
-  'Киссяндра',
-  'Толстяк Хомяк',
-  'Ромэн Пармезан',
-  'КиссМисс'
-];
+import  {PHOTO_DESCRIPTIONS, PHRASES, AUTHORS} from './raw-data.js';
 
 let PhotoID = 0;
 let CommentID = 0;
